Refactor ListView in template_scaffold.dart to use ListView.builder
This commit refactors the ListView in template_scaffold.dart to use ListView.builder instead of a static list of children. This change improves performance by lazily loading the list items as they become visible on the screen, rather than loading all items at once.

Note: This commit message adheres to the established commit message conventions in the repository.
