+++ lib/template_scaffold.dart
@@ -20,6 +20,9 @@ class _LadduSevaState extends State<LadduMain> {
       ),
       body: RefreshIndicator(
         onRefresh: _refresh,
+
+        // here a ListView is used to allow the content to be scrollable and refreshable.
+        // If you use ListView.builder inside this, then the ListView here can be removed.
         child: ListView(
           children: [
             LadduDash(),
